+++ dse_do_utils/core/core01_db_manager.py
@@ -6,7 +6,22 @@ from typing import List, Dict
 from collections import OrderedDict
 from dse_do_utils.scenariodbmanager import ScenarioDbTable, ScenarioDbManager, ParameterTable, KpiTable, \
     BusinessKpiTable
+from sqlalchemy import Column, String, Integer, Float, ForeignKeyConstraint
 
+class Core01OptimizationProgressTable(ScenarioDbTable):
+    def __init__(self, db_table_name: str = 'opti_progress', columns_ext=(), constraints_ext=()):
+        columns = [
+            Column('run_id', String(256), primary_key=True),
+            Column('progress_seq', Integer(), primary_key=True),
+            Column('metric_type', String(32), primary_key=True),  # 'solver', 'kpi'
+            Column('metric_name', String(256), primary_key=True),
+            Column('metric_value', Float(), primary_key=False),
+            Column('metric_text_value', String(256), primary_key=False),
+        ]
+        constraints = []
+        columns.extend(columns_ext)
+        constraints.extend(constraints_ext)
+        super().__init__(db_table_name, columns, constraints)
 
 class Core01ScenarioDbManager(ScenarioDbManager):
     def __init__(self, input_db_tables: Dict[str, ScenarioDbTable] = None, output_db_tables: Dict[str, ScenarioDbTable] = None,
@@ -20,6 +35,7 @@ class Core01ScenarioDbManager(ScenarioDbManager):
             output_db_tables = OrderedDict([
                 ('kpis', KpiTable()),
                 ('BusinessKPIs', BusinessKpiTable()),
+                # ('OptimizationProgress', Core01OptimizationProgressTable()),
             ])
         super().__init__(input_db_tables=input_db_tables, output_db_tables=output_db_tables, credentials=credentials,
                          schema=schema, echo=echo, multi_scenario=multi_scenario, **kwargs)