+++ dse_do_utils/scenariomanager.py
@@ -86,7 +86,7 @@ class ScenarioManager(object):
                  local_root: Optional[Union[str, Path]] = None, project_id: Optional[str] = None, project_access_token: Optional[str] = None, project=None,
                  template_scenario_name: Optional[str] = None, platform: Optional[Platform] = None,
                  inputs: Inputs = None, outputs: Outputs = None,
-                 local_relative_data_path: str = 'assets/data_asset', data_directory: str = None):
+                 local_relative_data_path: str = 'assets/data_asset', data_directory: Optional[Union[str, Path]] = None):
         """Create a ScenarioManager.
 
         Template_scenario_name: name of a scenario with an (empty but) valid model that has been successfully run at least once.
@@ -867,6 +867,7 @@ class ScenarioManager(object):
                 sheet_name = ScenarioManager._create_unique_abbreviated_name(table_name, 31,
                                                                              sheet_names)
                 sheet_names.add(sheet_name)
+                df = ScenarioManager.remove_timezone_from_datetime_columns(df)  # Remove timezone from datetime columns
                 df.to_excel(writer, sheet_name, index=False)
                 # Store row in table_index
                 table_index.append({'table_name': table_name, 'sheet_name': sheet_name, 'category': 'input'})
@@ -876,6 +877,7 @@ class ScenarioManager(object):
                 sheet_name = ScenarioManager._create_unique_abbreviated_name(table_name, 31,
                                                                              sheet_names)
                 sheet_names.add(sheet_name)
+                df = ScenarioManager.remove_timezone_from_datetime_columns(df)  # Remove timezone from datetime columns
                 df.to_excel(writer, sheet_name, index=False)
                 # Store row in table_index
                 table_index.append({'table_name': table_name, 'sheet_name': sheet_name, 'category': 'output'})
@@ -885,6 +887,20 @@ class ScenarioManager(object):
             index_df = pd.DataFrame(table_index)
             index_df.to_excel(writer, table_index_sheet, index=False)
 
+    @staticmethod
+    def remove_timezone_from_datetime_columns(df: pd.DataFrame) -> pd.DataFrame:
+        """Remove timezone from all datetime columns in a DataFrame.
+        This is necessary for the Excel writer, which does not support timezone-aware datetime columns.
+
+        Args:
+            df (pd.DataFrame): DataFrame with datetime columns
+
+        Returns:
+            pd.DataFrame: DataFrame with timezone removed from datetime columns
+        """
+        for col in df.select_dtypes(include=['datetime64[ns, UTC]']).columns:
+            df[col] = df[col].dt.tz_localize(None)
+        return df
     # -----------------------------------------------------------------
     # Load data from csv
     # -----------------------------------------------------------------